kernel: beginings of driver for virtio block devices

+++ sys/sys/types.h
@@ -43,7 +43,7 @@ typedef unsigned int ino_t;
 typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
 typedef unsigned int uint32_t;
-typedef unsigned long int uint64_t;
+typedef unsigned long long int uint64_t;
 typedef unsigned long int time_t;
 
 typedef int intptr_t;
